requirements.txt
pillow
gradio_client
keyboard
easyocr
pyttsx3
psutil